+++ src/object.h
@@ -82,8 +82,8 @@ void object_inherit_boundary(struct object* cell, const struct object* othercell
 struct vector* object_get_boundary(const struct object* cell);
 
 // ports
-void object_add_port(struct object* cell, const char* name, const struct generics* layer, const point_t* where, int storeanchor, double sizehint);
-void object_add_bus_port(struct object* cell, const char* name, const struct generics* layer, const point_t* where, int startindex, int endindex, unsigned int xpitch, unsigned int ypitch, int storeanchor, double sizehint);
+void object_add_port(struct object* cell, const char* name, const struct generics* layer, const point_t* where, int storeanchor, unsigned int sizehint);
+void object_add_bus_port(struct object* cell, const char* name, const struct generics* layer, const point_t* where, int startindex, int endindex, unsigned int xpitch, unsigned int ypitch, int storeanchor, unsigned int sizehint);
 const struct vector* object_get_ports(const struct object* cell);
 
 // alignment box and bounding box
@@ -188,7 +188,7 @@ struct port_iterator;
 struct port_iterator* object_create_port_iterator(const struct object* cell);
 int port_iterator_is_valid(struct port_iterator* it);
 void port_iterator_next(struct port_iterator* it);
-void port_iterator_get(struct port_iterator* it, const char** portname, const point_t** portwhere, const struct generics** portlayer, int* portisbusport, int* portbusindex, double* sizehint);
+void port_iterator_get(struct port_iterator* it, const char** portname, const point_t** portwhere, const struct generics** portlayer, int* portisbusport, int* portbusindex, unsigned int* sizehint);
 void port_iterator_destroy(struct port_iterator* it);
 
 #endif // OPC_OBJECT_H
